requirements.txt
PyQt5>=5.15.9
cubiomespi>=0.1.0
numpy>=1.24.3
pyinstaller>=5.13.0
